fuse-adaptor:Add splice api for all loop mode
This patch adds splice I/O support to the fuse request receive process for all
three loop modes. Due to some invisible struct definitions in the libfuse code,
parts of the libfuse interfaces have to be rewritten, which are all placed in
fuse-compat.cpp. For the same reason, we cannot automatically switch between splice(2)
and read(2) as libfuse does internally, so currently only the forced splice method
and the normal read interface are supported.