Fix cross-compiling by using CMAKE_SYSTEM_PROCESSOR and respecting -march|-mcpu (#904)
* Fix cross-compiling by using CMAKE_SYSTEM_PROCESSOR rather than uname -m

There is no need to check the architecture twice. Also check for arm64
(MacOS) wherever we check for aarch64 (Linux).

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>

* Respect aarch64 -march=|-mcpu= flag and check whether -mcpu=native works

Packagers may specify a target system up front, which should be
respected. The CRC extensions are not strictly necessary as there is a
software fallback.

Otherwise, use -mcpu=native if it works (it won't when cross-compiling)
or -mcpu=generic+crc.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>

---------

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>